move tooltip to dedicated directory
