+++ components/SegmentManagement.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react"
 
 import Button from "@components/Button/Button/Button"
-import RequestTooltip from "@components/RequestTooltip"
+import RequestTooltip from "@components/Tooltip/RequestTooltip"
 import SegmentCreationWizard from "@components/SegmentCreationWizard"
 import { SegmentObject } from "@components/SegmentObject"
 import { useSegmentManagementLink } from "@hooks/use-segment-management-link"
