+++ components/SyncObject.js
@@ -2,7 +2,7 @@ import { useState } from "react"
 
 import Button from "@components/Button/Button/Button"
 import Card from "@components/Card/Card"
-import RequestTooltip from "@components/RequestTooltip"
+import RequestTooltip from "@components/Tooltip/RequestTooltip"
 import SyncEditWizard from "@components/SyncEditWizard"
 import { SyncStatus } from "@components/SyncStatus"
 import Toggle from "@components/Toggle/Toggle"
